+++ vptree.h
@@ -238,9 +238,9 @@ private:
 
         // If current node within radius tau
         if(dist < _tau) {
-            if(heap.size() == k) heap.pop();                 // remove furthest node from result list (if we already have k results)
+            if((int)heap.size() == k) heap.pop();                 // remove furthest node from result list (if we already have k results)
             heap.push(HeapItem(node->index, dist));           // add current node to result list
-            if(heap.size() == k) _tau = heap.top().dist;     // update value of tau (farthest point in result list)
+            if((int)heap.size() == k) _tau = heap.top().dist;     // update value of tau (farthest point in result list)
         }
 
         // Return if we arrived at a leaf
